Rename mass univariate testing utils
